# docs/swagger/swagger.yaml
definitions:
...
  http.standardJsonResponse:
    properties:
      data:
        type: object
      error:
        type: string
      meta:
        type: object
      success:
        type: boolean
    type: object
  http.storyboardColumnAddRequestBody:
    properties:
      goalId:
        type: string
    required:
    - goalId
    type: object
  http.storyboardCreateRequestBody:
    properties:
      facilitatorCode:
        type: string
      joinCode:
        type: string
      storyboardName:
        type: string
    required:
    - storyboardName
    type: object
  http.storyboardGoalAddRequestBody:
    properties:
      name:
        minLength: 1
        type: string
    required:
    - name
    type: object
  http.storyboardStoryAddRequestBody:
    properties:
      columnId:
        type: string
      goalId:
        type: string
    required:
    - columnId
    - goalId
    type: object
  http.storyboardStoryMoveRequestBody:
    properties:
      columnId:
        type: string
      goalId:
        type: string
      placeBefore:
        type: string
    required:
    - columnId
    - goalId
    type: object
...
  thunderdome.Storyboard:
    properties:
      color_legend:
        items:
          $ref: '#/definitions/thunderdome.Color'
        type: array
      createdDate:
        type: string
      facilitatorCode:
        type: string
      facilitators:
        items:
          type: string
        type: array
      goals:
        items:
          $ref: '#/definitions/thunderdome.StoryboardGoal'
        type: array
      id:
        type: string
      joinCode:
        type: string
      name:
        type: string
      owner_id:
        type: string
      personas:
        items:
          $ref: '#/definitions/thunderdome.StoryboardPersona'
        type: array
      teamId:
        type: string
      teamName:
        type: string
      updatedDate:
        type: string
      users:
        items:
          $ref: '#/definitions/thunderdome.StoryboardUser'
        type: array
    type: object
...
info:
  contact:
    name: Steven Weathers
    url: https://github.com/StevenWeathers/thunderdome-planning-poker
  description: |-
    Thunderdome Planning Poker API for both Internal and External use.
    WARNING: Currently not considered stable and is subject to change until 1.0 is released.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thunderdome API
  version: BETA
paths:
...
  /{orgId}/departments/{departmentId}/teams/{teamId}/users/{userId}/storyboards:
    post:
      description: Create a storyboard associated to the user
      parameters:
      - description: the user ID
        in: path
        name: userId
        required: true
        type: string
      - description: the organization ID
        in: path
        name: orgId
        type: string
      - description: the department ID
        in: path
        name: departmentId
        type: string
      - description: the team ID
        in: path
        name: teamId
        type: string
      - description: new storyboard object
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardCreateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Storyboard'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Storyboard
      tags:
      - storyboard
...
  /{orgId}/teams/{teamId}/users/{userId}/storyboards:
    post:
      description: Create a storyboard associated to the user
      parameters:
      - description: the user ID
        in: path
        name: userId
        required: true
        type: string
      - description: the organization ID
        in: path
        name: orgId
        type: string
      - description: the team ID
        in: path
        name: teamId
        type: string
      - description: new storyboard object
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardCreateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Storyboard'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Storyboard
      tags:
      - storyboard
...
  /storyboards:
    get:
      description: get list of storyboards
      parameters:
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      - description: Only active storyboards
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Storyboard'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Storyboards
      tags:
      - storyboard
  /storyboards/{storyboardId}:
    delete:
      description: Delete a storyboard
      parameters:
      - description: the storyboard ID
        in: path
        name: storyboardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Storyboard Delete
      tags:
      - storyboard
    get:
      description: get storyboard by ID
      parameters:
      - description: the storyboard ID to get
        in: path
        name: storyboardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Storyboard'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Storyboard
      tags:
      - storyboard
  /storyboards/{storyboardId}/columns:
    post:
      description: Add a column to a storyboard goal
      parameters:
      - description: the storyboard ID
        in: path
        name: storyboardId
        required: true
        type: string
      - description: request body for adding a column
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardColumnAddRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Storyboard Column Add
      tags:
      - storyboard
  /storyboards/{storyboardId}/goals:
    post:
      description: Add a goal to a storyboard
      parameters:
      - description: the storyboard ID
        in: path
        name: storyboardId
        required: true
        type: string
      - description: the goal to add
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardGoalAddRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Storyboard Goal Add
      tags:
      - storyboard
  /storyboards/{storyboardId}/stories:
    post:
      description: Add a story to a storyboard goal column
      parameters:
      - description: the storyboard ID
        in: path
        name: storyboardId
        required: true
        type: string
      - description: request body for adding a story
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardStoryAddRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Storyboard Story Add
      tags:
      - storyboard
  /storyboards/{storyboardId}/stories/{storyId}/move:
    put:
      description: Move a story in a storyboard
      parameters:
      - description: the storyboard ID
        in: path
        name: storyboardId
        required: true
        type: string
      - description: the story ID
        in: path
        name: storyId
        required: true
        type: string
      - description: target goal column and place before story
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardStoryMoveRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Storyboard Story Move
      tags:
      - storyboard
...
  /teams/{teamId}/users/{userId}/storyboards:
    post:
      description: Create a storyboard associated to the user
      parameters:
      - description: the user ID
        in: path
        name: userId
        required: true
        type: string
      - description: the team ID
        in: path
        name: teamId
        type: string
      - description: new storyboard object
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardCreateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Storyboard'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Storyboard
      tags:
      - storyboard
...
  /users/{userId}/storyboards:
    get:
      description: get list of storyboards for the user
      parameters:
      - description: the user ID to get storyboards for
        in: path
        name: userId
        required: true
        type: string
      - description: Max number of results to return
        in: query
        name: limit
        type: integer
      - description: Starting point to return rows from, should be multiplied by limit
          or 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.Storyboard'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Storyboards
      tags:
      - storyboard
    post:
      description: Create a storyboard associated to the user
      parameters:
      - description: the user ID
        in: path
        name: userId
        required: true
        type: string
      - description: new storyboard object
        in: body
        name: storyboard
        schema:
          $ref: '#/definitions/http.storyboardCreateRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.Storyboard'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Storyboard
      tags:
      - storyboard
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"